Dylan Lucas - Figured out how to do a List method with the playerBag so I can add and remove objects from the players Inventory. Took this method and got the take method working properly and added the remove method which is now working properly.
Need to add more objects per room, which would then create more objects of the Bag list, to get the world filled with more items, but this is a good starting point to work off of.

Need to add the "Open" method for any possible doors or chests I believe. Most likely will leave that for chests cause in this world everybody trusts everybody and doors are always unlocked! Chests on the other hand are not unlocked.